Moviendo_publico/Moviendo-main/back/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
drf-spectacular==0.28.0
djangorestframework-camel-case==1.4.2
requests==2.32.3
